pyproject.toml
[project]
name = "python-service"
version = "0.1.0"
description = "Add your description here"
license = "MIT"
requires-python = ">=3.13"

dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "faststream[kafka]>=0.5.48",
    "litestar-granian>=0.14.2",
    "litestar[jwt,pydantic,sqlalchemy]>=2.17.0",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.9",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 130
exclude = ["migrations"]

[tool.pytest.ini_options]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="function"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
    "B904",   # Allow raising exceptions without from e, for HTTPException
]
